Create new instruction models (#30)
Added two new types of models to represent data in the system.
The data_size models represent raw data that doesn't have an intended purpose in the system. They merely model a number of bits and provide some basic logic functions.
Sizes are Word (16), Byte(8), and Nibble(4).
The data_type models are extensions of the data_size models and have a particular purpose in the system and so their added functionality is tailored to that.
(Most of these models are actually pretty similar and there is not much difference between them at the moment, they are mostly used to make the system a little more readable and set it up to make logging easy).

The rest of the PR is refactoring the system to use these models.