Creating Interfaces for Instructions to Implements (#35)
SInce there was a decent amount of reproduced logic in a lot of the instructions, I figured the best approach to eliminate some of that would be to make the instructions a little more meta-data driven. I made an interface that contained keys for often used pieces of data like the number of bytes an instruction uses and cycles an instruction takes. This way a single function can do a lot of the logic that was being reproduced.

In this PR I:
* Made a new Interface for all instructions to implement
* Updated all instructions to be objects that implement said interface
* Added action functions for all instructions that have unique logic in them (most)
* Created a new function in the Z80 class that replicates what jsGB did for executing functions, but now uses the meta-data functions instead
* Removed old logic from run function and am now calling new Z80 function instead
* Added return types to Z80 functions