Adding integration testing (#73)
This PR adds Jest and some extra dependencies it needs to handle Typescript, Canvas, and EcmaScript modules. It adds a single test who's job is to set up a mock html page to run and then load the test program into the emulator and run the actual emulator. I had to tweak a few minor things in jsGB in order to get things to work, but nothing too invasive. Changes are as follows:

* Forgot to actually subtract the value from the A register

* Adding Jest and its necessary dependancies

* Exporting handler function so test can load a rom into the emulator

* Updating typescript config so jest will run

* Adding test rom so tests can load it

* Adding configuration for jest

* Adding test for running the emulator

* adding canvas package for jsdom

* removing unnecessary reset

* Adding sleep functions to allow emulator load rom and run. Also added needed elements to mock page so the emulator will run correctly