Adding flag reset scheme (#76)
This PR provides a new schema for resetting flags at the beginning of every instruction execution. Basically, there are now properties for each flag on the instruction meta-data that will determine what the flags should be set to before the instruction is executed. A 1 or 0 will just set the flag to either value, but a question mark will set the flag to 0 in anticipation that there is conditional logic in the instruction for setting the flag to 1. This just makes it a little simpler to set flags to static values and keeps the instruction logic becoming a big chain of if/else statements.

* Adding logic to in execution to read metadata for flag states

* Adding flag reset keys to interface

* Adding flag reset states to 8bitArithmetic functions

* Fixing issue with context. When pulling a function from an object and storing the reference in a variable, the function loses the object from context

* Adding flag resets to all 8 bit arithmetic functions

* Adding flag reset parameters to all bit operation instructions