Fixing Major GPU Bug and some Minor Timing issues and Semantic Bugs (#90)
There were a couple of things contributing to the the GPU timing issues. The main two were that not all instructions had the correct clock cycles and the GPU timing mechanism wasn't taking into account overflow cycles on reset. This was causing the frame to end on the wrong instruction, and the GPU stages were totally out of sync with the actual CPU timer. To fix this I:

* Fixed timing for certain instructions

* Fixing logic for BIT functions

* Fixing starting mode and resetting the modeClock with leftover cycles instead of 0

* Removing Logging and Debugging Temp Code (Too speed up debugging)